// src/app/dyno-admin/models/IEmployer.ts
export interface IEmployer{
    Name: string;
    Password: string;
    Email: string;
    Role: string;
    walletpublickey:string;
    numeroTelEntreprise: string;
    adresseFacturation: string;
    codeTVA: string;
    adresseEntreprise: string;
    EmailRH : string;
    NumTelRH: string;
    matriculeFiscale:string;
    paymentMethode: string;
}